refactor: Annotation 관련 확장 함수 선언
