+++ di/src/main/java/com/android/di/annotation/AnnotationBinders.kt
@@ -1,0 +1,14 @@
+package com.android.di.annotation
+
+import kotlin.reflect.KClass
+
+fun List<Annotation>.hasQualifier(): Boolean =
+    any { it.isQualifier() }
+
+fun List<Annotation>.qualifierAnnotation(): KClass<out Annotation>? {
+    return firstOrNull { it.isQualifier() }?.annotationClass
+}
+
+fun Annotation.isQualifier(): Boolean {
+    return annotationClass.annotations.any { it.annotationClass == Qualifier::class }
+}
